Map `long long` IDL type to Swift `Int32`
This fixes an issue with `func vertexAttribPointer(index: GLuint, size: GLint, type: GLenum, normalized: GLboolean, stride: GLsizei, offset: GLintptr)` passing `offset` value of `BigInt` type to the JS function due to `public typealias GLintptr = Int64`, which led to `can't cast BigInt to number` runtime errors.

Either WebGL spec is wrong, or browsers implement it incorrectly. Rust folks had a similar issue and they went with `i32`.

+++ Sources/WebIDLToSwift/WebIDL+SwiftRepresentation.swift
@@ -541,7 +541,7 @@ extension IDLType: SwiftRepresentable {
         "byte": "Int8",
         "short": "Int16",
         "long": "Int32",
-        "long long": "Int64",
+        "long long": "Int32",
         "Function": "JSFunction",
         "bigint": "__UNSUPPORTED_BIGINT__",
     ]
